.gitlab/ci/scripts/requirements.txt
PyYAML
python-gitlab
requests
envparse
cryptography
wcmatch
